+++ .trebired/code-discipline/config.ts
@@ -2,6 +2,6 @@ import { defineConfig } from "@trebired/code-discipline";
 
 export default defineConfig({
   presets: {
-    use: ["@trebired/code-discipline-config"],
+    use: ["@trebired/configs"],
   },
 });